# common-platform-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/simulations/status/{status}:
    get:
      tags:
        - Simulations
      summary: Get simulations by status
      description: Returns all simulations with a specific status
      operationId: getSimulationsByStatus
      parameters:
        - name: status
          in: path
          description: Status of the simulations
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved simulations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '500':
          description: Internal server error
  /api/v1/simulations/country/{countryId}:
    get:
      tags:
        - Simulations
      summary: Get simulations by country
      description: Returns all simulations for distributors operating in a specific country
      operationId: getSimulationsByCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved simulations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '500':
          description: Internal server error
  /api/v1/simulations/application/{applicationId}:
    get:
      tags:
        - Simulations
      summary: Get simulation by application ID
      description: Returns a simulation based on its application ID
      operationId: getSimulationByApplicationId
      parameters:
        - name: applicationId
          in: path
          description: ID of the application
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved simulation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '404':
          description: Simulation not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/simulations/status/{status}:
    get:
      tags:
        - Distributor Simulations
      summary: Get simulations by status
      description: Returns simulations with a specific status for a distributor
      operationId: getSimulationsByDistributorIdAndStatus
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: path
          description: Status of the simulations
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor simulations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '500':
          description: Internal server error
...
components:
  schemas:
...
    DistributorSimulationDTO:
      required:
        - applicationId
        - distributorId
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        distributorId:
          type: integer
          format: int64
        applicationId:
          type: integer
          format: int64
        simulationStatus:
          maxLength: 50
          minLength: 0
          type: string
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          format: int64
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: integer
          format: int64
          readOnly: true
...